WIP
The main idea to collect connections for swift driver.
Maybe we need to make context manager for this operations
to be sure that after the end (whether it good or bad) of
operation we will delete connection from list.

Now tested with swift, that at least run_backup command works

Change-Id: I882cf58b9e01cdfd30afd8466a6fe1b4cddc553a